finito e ordinato tutti i test

+++ src/main/java/it/prova/gestioneordiniarticolicategorie/service/OrdineServiceImpl.java
@@ -6,12 +6,14 @@ import javax.persistence.EntityManager;
 
 import it.prova.gestioneordiniarticolicategorie.dao.EntityManagerUtil;
 import it.prova.gestioneordiniarticolicategorie.dao.OrdineDAO;
+import it.prova.gestioneordiniarticolicategorie.exception.OrdineConArticoliException;
+import it.prova.gestioneordiniarticolicategorie.model.Categoria;
 import it.prova.gestioneordiniarticolicategorie.model.Ordine;
 
 public class OrdineServiceImpl implements OrdineService {
 
 	private OrdineDAO ordineDAO;
-	
+
 	@Override
 	public List<Ordine> listAll() throws Exception {
 		EntityManager entityManager = EntityManagerUtil.getEntityManager();
@@ -71,7 +73,7 @@ public class OrdineServiceImpl implements OrdineService {
 		} finally {
 			EntityManagerUtil.closeEntityManager(entityManager);
 		}
-		
+
 	}
 
 	@Override
@@ -104,6 +106,9 @@ public class OrdineServiceImpl implements OrdineService {
 
 			ordineDAO.setEntityManager(entityManager);
 			entityManager.getTransaction().begin();
+			if (!ordineDAO.get(id).getArticoli().isEmpty()) {
+				throw new OrdineConArticoliException("Ordine con articoli");
+			}
 			ordineDAO.delete(ordineDAO.get(id));
 			entityManager.getTransaction().commit();
 
@@ -122,7 +127,146 @@ public class OrdineServiceImpl implements OrdineService {
 	public void setOrdineDAO(OrdineDAO ordineDAO) {
 
 		this.ordineDAO = ordineDAO;
-		
+
+	}
+
+	@Override
+	public List<Ordine> ordiniDiUnaCategoria(Categoria categoria) throws Exception {
+		EntityManager entityManager = EntityManagerUtil.getEntityManager();
+
+		if (categoria.getId() == null) {
+			throw new RuntimeException("Errore input");
+		}
+
+		try {
+
+			ordineDAO.setEntityManager(entityManager);
+
+			return ordineDAO.ordiniDiUnaCategoria(categoria.getId());
+
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw e;
+		} finally {
+			EntityManagerUtil.closeEntityManager(entityManager);
+
+		}
+	}
+
+	@Override
+	public List<Categoria> distinteCategorieDiUnOrdine(Ordine ordine) throws Exception {
+		EntityManager entityManager = EntityManagerUtil.getEntityManager();
+
+		if (ordine.getId() == null) {
+			throw new RuntimeException("Errore input");
+		}
+
+		try {
+
+			ordineDAO.setEntityManager(entityManager);
+
+			return ordineDAO.distinteCategorieDiUnOrdine(ordine.getId());
+
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw e;
+		} finally {
+			EntityManagerUtil.closeEntityManager(entityManager);
+
+		}
+	}
+
+	@Override
+	public Ordine caricaOrdinePiuRecenteDiUnaCategoria(Categoria categoria) throws Exception {
+		EntityManager entityManager = EntityManagerUtil.getEntityManager();
+
+		if (categoria.getId() == null) {
+			throw new RuntimeException("Errore input");
+		}
+
+		try {
+
+			ordineDAO.setEntityManager(entityManager);
+
+			return ordineDAO.getOrdinePiuRecenteByCateoria(categoria.getId());
+
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw e;
+		} finally {
+			EntityManagerUtil.closeEntityManager(entityManager);
+
+		}
+	}
+
+	@Override
+	public List<String> caricaCodiciDiCategorieDiArtcioliDiOrdiniEffettuatiInMeseEAnno(int mese, int anno)
+			throws Exception {
+		EntityManager entityManager = EntityManagerUtil.getEntityManager();
+
+		if (mese > 12 || mese < 1 || anno < 1000 || anno > 2999) {
+			throw new RuntimeException("Errore input");
+		}
+
+		try {
+
+			ordineDAO.setEntityManager(entityManager);
+
+			return ordineDAO.getCodiciDiCategorieDiMeseEAnno(mese, anno);
+
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw e;
+		} finally {
+			EntityManagerUtil.closeEntityManager(entityManager);
+
+		}
+	}
+
+	@Override
+	public Double caricaSommaPrezziDiUnDesinatario(String nomeDestinatario) throws Exception {
+		EntityManager entityManager = EntityManagerUtil.getEntityManager();
+
+		if (nomeDestinatario == null) {
+			throw new RuntimeException("Errore input");
+		}
+
+		try {
+
+			ordineDAO.setEntityManager(entityManager);
+
+			return ordineDAO.getSumPrezziDiDesinatario(nomeDestinatario);
+
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw e;
+		} finally {
+			EntityManagerUtil.closeEntityManager(entityManager);
+
+		}
+	}
+
+	@Override
+	public List<String> listaDiIndirizziDiordiniAventiArticoliConSeriale(String segmentoSeriale) throws Exception {
+		EntityManager entityManager = EntityManagerUtil.getEntityManager();
+
+		if (segmentoSeriale == null) {
+			throw new RuntimeException("Errore input");
+		}
+
+		try {
+
+			ordineDAO.setEntityManager(entityManager);
+
+			return ordineDAO.listaDiIndirizziDiordiniAventiArticoliConSeriale(segmentoSeriale);
+
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw e;
+		} finally {
+			EntityManagerUtil.closeEntityManager(entityManager);
+
+		}
 	}
 
 }
